Backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.116.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msrest==0.7.1
multidict==6.6.4
narwhals==2.2.0
nest-asyncio==1.6.0
numpy==2.2.6
oauthlib==3.3.1
openai==1.102.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.4
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
streamlit==1.49.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.4.0
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
yarl==1.20.1